+++ includes/autoloader.inc.php
@@ -48,7 +48,7 @@ function __autoload($class_name)
         }
     }
 
-    $error = Error::singleton();
+    $error = PSI_Error::singleton();
 
     $error->addError("_autoload(\"".$class_name."\")", "autoloading of class file (class.".$class_name.".inc.php) failed!");
     $error->errorsAsXML();
@@ -66,7 +66,7 @@ function __autoload($class_name)
  */
 function errorHandlerPsi($level, $message, $file, $line)
 {
-    $error = Error::singleton();
+    $error = PSI_Error::singleton();
     $error->addPhpError("errorHandlerPsi : ", "Level : ".$level." Message : ".$message." File : ".$file." Line : ".$line);
 }